+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - 添加 Valid 方法，用于验证文档是否正确；
-- 为 Request 和 Param 添加了四个用于描述 XML 数据的字段；
+- 为 Request 和 Param 添加了五个用于描述 XML 数据的字段；
 - 为 apidoc 元素添加 mimetype，用于指定全局可用的 mimetype 值；
 
 ### Changed